backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "braintrust>=0.0.160",
    "fastapi>=0.135.2",
    "httpx>=0.28.0",
    "openapi-spec-validator>=0.7.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.2",
    "restate-sdk>=0.16.0",
    "uvicorn>=0.42.0",
]

[project.scripts]
search-cli = "backend.manual_cli:main"
